chore: rename vendor package to vendorservice
