+++ backend/internal/vendorservice/url_builder_interface.go
@@ -1,6 +1,6 @@
 package vendor
 
-//go:generate mockgen -destination=../mock/vendor/url_builder.go -package=mockvendor . BookURLBuilder
+//go:generate mockgen -destination=../mock/vendorservice/url_builder.go -package=mockvendorservice . BookURLBuilder
 type BookURLBuilder interface {
 	BookURL(bookID string) string
 	ChapterListURL(bookID string) string
